Move OptionsMetadata derive tests to trait crate
