+++ crates/ruff_options_metadata/Cargo.toml
@@ -13,6 +13,11 @@ license = { workspace = true }
 [dependencies]
 serde = { workspace = true, optional = true }
 
+[dev-dependencies]
+ruff_macros = { workspace = true }
+
+serde = { workspace = true, features = ["derive"] }
+
 [lints]
 workspace = true
